Выполнил 2 пункт задания
