+++ learning-center-api/pom.xml
@@ -5,8 +5,8 @@
     <modelVersion>4.0.0</modelVersion>
     <parent>
         <groupId>space.irsi7</groupId>
-        <artifactId>StudentsEducationSystem</artifactId>
-        <version>1.0-SNAPSHOT</version>
+        <artifactId>learning-center</artifactId>
+        <version>1.0.0-SNAPSHOT</version>
     </parent>
 
     <artifactId>learning-center-api</artifactId>
