+++ learning-center-impl/src/main/java/space/irsi7/app/MyAppContextConfiguration.java
@@ -6,18 +6,18 @@ import org.springframework.context.annotation.Configuration;
 import space.irsi7.bpp.InjectRandomMarksBeanPostProcessor;
 import space.irsi7.models.Student;
 
-@Configuration
-@ComponentScan(basePackages = "space.irsi7")
-public class MyAppContextConfiguration {  // (1)
-
-    @Bean
-    public InjectRandomMarksBeanPostProcessor eventBusBeanPostProcessor() {
-        return new InjectRandomMarksBeanPostProcessor();
-    }
-
-    @Bean
-    public Student studentBean(){
-        return new Student();
-    }
-
-}
+//@Configuration
+//@ComponentScan(basePackages = "space.irsi7")
+//public class MyAppContextConfiguration {  // (1)
+//
+//    @Bean
+//    public InjectRandomMarksBeanPostProcessor eventBusBeanPostProcessor() {
+//        return new InjectRandomMarksBeanPostProcessor();
+//    }
+//
+//    @Bean
+//    public Student studentBean(){
+//        return new Student();
+//    }
+//
+//}
