--- learning-center-impl/src/main/resources/logback.xml
@@ -1,37 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<configuration>
-    <property name="LOG_ROOT" value="logs" />
-    <property name="LOG_FILE_NAME" value="app-log" />
-
-    <appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
-        <encoder>
-            <pattern>%d{HH:mm:ss.SSS} [%thread] %-5level %logger{36} - %msg%n</pattern>
-        </encoder>
-    </appender>
-
-    <appender name="FILE" class="ch.qos.logback.core.rolling.RollingFileAppender">
-        <file>${LOG_ROOT}/${LOG_FILE_NAME}.log</file>
-        <rollingPolicy class="ch.qos.logback.core.rolling.SizeAndTimeBasedRollingPolicy">
-            <fileNamePattern>${LOG_ROOT}/${LOG_FILE_NAME}-%d{yyyy-MM-dd}.%i.log</fileNamePattern>
-            <!-- each archived file's size will be max 10MB -->
-            <maxFileSize>10MB</maxFileSize>
-            <!-- 30 days to keep -->
-            <maxHistory>1</maxHistory>
-            <!-- total size of all archive files, if total size > 100GB, it will delete old archived file -->
-            <totalSizeCap>1GB</totalSizeCap>
-        </rollingPolicy>
-        <encoder>
-            <pattern>%d{HH:mm:ss.SSS} [%thread] %-5level %logger{36} - %msg%n</pattern>
-        </encoder>
-    </appender>
-
-    <logger name="space.irsi7" level="INFO" additivity="false">
-        <appender-ref ref="STDOUT" />
-        <appender-ref ref="FILE"/>
-    </logger>
-
-    <root level="info">
-        <appender-ref ref="STDOUT" />
-        <appender-ref ref="FILE" />
-    </root>
-</configuration>
